requirements.txt
asgiref==3.5.0
black @ git+https://github.com/psf/black@147526451a43e9296ca963ed8e6b7224db93e69b
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==4.0.2
django-allauth==0.47.0
django-treebeard==4.5.1
djangorestframework==3.13.1
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
